Removed broken IP validation
Technically, the IP validation worked flawlessly. But we don't want to
limit people to IPs. Any host should be valid (eg, "localhost"). Somehow
it let this through in some cases, anyway... Now there's just no validation,
which is just fine.

Also having the default text be "localhost" for ease of development.
This should be changed for production.