level selection
The level select menu has been added
When a level is selected the players are spawned into the spawnpoints
A added some basic movement to the players, this works in multiplayer.
Each player is controlled by one client.
I didn't commit this to the main branch because the character selection
menu spawns players for some reason.